pyproject.toml
[tool.poetry]
name = "flask-account"
version = "0.1.0"
description = ""
authors = ["Tryd0g0lik <work80@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
flask = {extras = ["async"], version = "^3.0.3"}
sqlalchemy = "^2.0.35"

[tool.poetry.group.dev.dependencies]
asyncio = "^3.4.3"
autohooks = "^24.2.0"
flake8 = "^7.1.1"
postgres = "4.0"
psycopg2-binary = "2.9.9"
pre-commit = "^3.8.0"
markdown = "^3.7"
pylint = "^3.3.1"
isort = "^5.13.2"
black = "^24.8.0"
flask-bootstrap = "^3.3.7.1"
flask-jwt-extended = "^4.6.0"
flask-swagger-ui = "^4.11.1"
uuid = "^1.30"
flask-wtf = "^1.2.1"
wtforms = "^3.1.2"
flask-sqlalchemy = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
